Fixed semantic of log blocks
The ssd.config defines the log blocks size to be used by the FTLs.
However, the name of those variables does not reflect the real
behavior of them, as well as the comments.
These variables are used as block-size unit, not as page-size
unit, so it causes missunderstanding during usage.

Signed-off-by: Cassiano Campes <cassianocampes@gmail.com>